fix: set job id to references channels collect

+++ src/data/entities/Metrics.ts
@@ -17,7 +17,7 @@ export class Metrics {
   @Column('text')
   channel_theme: string | undefined;
 
-  @Column('int')
+  @Column('float')
   score: number | undefined;
 
   @Column('int')
